Code/main/mpquic-rl/requirements.txt
absl-py>=0.9.0
astor>=0.8.1
astroid>=2.3.3
cycler>=0.10.0
gast>=0.3.3
google-pasta>=0.2.0
grpcio>=1.27.2
h5py>=2.10.0
isort>=4.3.21
joblib>=0.14.1
Keras-Applications>=1.0.8
Keras-Preprocessing>=1.1.0
kiwisolver>=1.1.0
lazy-object-proxy>=1.4.3
Markdown>=3.2.1
matplotlib>=3.2.1
mccabe>=0.6.1
numpy>=1.18.4
Pillow>=7.0.0
# pkg-resources>=0.0.0
protobuf>=3.11.3
pylint>=2.4.4
pyparsing>=2.4.6
python-dateutil>=2.8.1
pyzmq>=19.0.0
scikit-learn>=0.23.0
scipy>=1.4.1
six>=1.14.0
tensorboard>=1.14.0
tensorflow>=1.14.0
tensorflow-estimator>=1.14.0
termcolor>=1.1.0
tflearn>=0.3.2
threadpoolctl>=2.0.0
typed-ast>=1.4.1
Werkzeug>=1.0.0
wrapt>=1.11.2
pandas

pexpect>=4.0.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch>=2.0.0
learn2learn # currently broken because of qpth, install qpth-0.0.15.tar.gz
gymnasium
cherry-rl
policies
